third_party/WebKit/LayoutTests/editing/deleting/delete-ligature-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<textarea>&#x0E27;&#x0E31;</textarea>
<script>
test(() => {
    assert_not_equals(window.eventSender, undefined,
        'This test requires window.eventSender');
    const textarea = document.querySelector('textarea');
    textarea.setSelectionRange(2, 2);
    textarea.focus();
    eventSender.keyDown('Backspace', null);
    assert_equals(textarea.value, '\u0E27');
}, 'Backspace key to delete a last character of ligature U+0E27 and U+0E31');
</script>
